Add files via upload
This is my web scraping project. Here I extract some information from a travel agency (www.booking.com) website. It gives you the hotel information where you want to go. In this project, I used selenium which is very useful for automation. That project is beneficial for me and now I am a master of selenium.